Green palette
Created an alternative version with only green shades used,
coherence throughout the website with a colour perceived as
soothing, and due to taking up the largest amount of visible spectrum
could be idea for a mono-chromatic version of the website. Also green
links well to the subject of the website.